+++ vite.config.ts
@@ -3,7 +3,7 @@ import react from '@vitejs/plugin-react'
 import { crx, type ManifestV3Export } from '@crxjs/vite-plugin'
 import { fileURLToPath, URL } from 'node:url'
 
-const VERSION = '0.3.0'
+const VERSION = '0.5.0'
 
 const manifest: ManifestV3Export = {
   manifest_version: 3,